Add tests for FRE to ZTE and guaranteed enrollment transitions
This adds a test for an FRE to ZTE transition when a device is freed
from FRE and is pre-provisioned for ZTE or guaranteed enrollment.

This CL also refactors the existing test for FRE to initial state
transition for license packaging, keeping those tests more comsistent
and close to each other.

See go/cros-oobe-state-determinations for design.

Bug: chromium:1102607
Test: xvfb-run unit_tests --gtest_filter="*AutoEnrollmentClientImplTest*"
Change-Id: I519f38884c31a028de0d7bf74be10bbc952c0289
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283240
Commit-Queue: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Roman Aleksandrov <raleksandrov@google.com>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786935}