Remove SecurityContext from Document, centralize SecurityContext on ExecutionContext
When ExecutionContext moved from Document to LocalDOMWindow, we left
SecurityContext on Document. Moving it seemed complicated, partly because
of the number of callsites that assumed a Document would always have
access to a valid SecurityContext, and partly because of the complexity
of handling the cases where a LocalDOMWindow is reused when navigating
away from the initial empty document.

The work in https://crbug.com/1101402 has gotten us to the point where
almost no code depends directly on Document::GetSecurityContext().
Callers instead go through ExecutionContext, and are already equipped
to handle the possibility that the ExecutionContext is gone. This CL
takes care of the last few cases, and adds a way to Reinitialize() a
SecurityContext for LocalDOMWindow reuse, allowing us to move the
SecurityContext off of Document.

Now that the special-case of Document holding a SecurityContext is
unnecessary, we no longer need to have the non-window ExecutionContext
subclasses hold their own SecurityContext. Instead, ExecutionContext
can own the SecurityContext.

Bug: 1101402
Change-Id: I84b5fc9a31b88f7b7cd455136e458704a1a83ef6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2240902
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786957}