SuggestionWindowView cleanup part 3: Eliminate SettingLinkView
This custom View simply contained a single Link in a FillLayout.  The
Link can just as easily be added directly to SuggestionWindowView,
eliminating the boilerplate of the class and wrapper View.

This also scoped a couple of related bits:
(1) The link click handler
(2) The code to highlight the link
(3) The link highlight tracking bool

While we could simply move these to equivalent locations on
SuggestionWindowView, it's possible to do better:
(1) can be replaced by a lambda.
(2) can be replaced by a helper function that also obviates the need for
    SetLearnMoreButtonHighlighted().  This helper also eliminates the
    previous SchedulePaint() calls, since View::SetBackground() does
    that automatically.
(3) can be eliminated entirely, since the presence of the View's
    background is equivalent to whether the View is highlighted, so the
    highlighting code can just test that.  It's slightly cryptic but
    cheaper and simpler, and the same reasoning lets us eliminate
    |is_learn_more_button_highlighted| (no trailing '_'!) as well.

Also wrapped the file-scope stuff in an anonymous namespace, which is
common (and arguably clearer) style in Chromium.

Bug: none
Change-Id: If174f45418fb35ed3059c09677dee9c28e243295
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2288848
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Keith Lee <keithlee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786969}