qr_code_generator: misc fixes
The total_bytes for a version can be computed rather than specified. The
static_asserts can be moved into the constructor for a constexpr
function using the fact that the compiler must be able to statically
prove that all calls to __builtin_unreachable are, in fact, unreachable.

Change-Id: I29ba54497173b55dc2a5e7873b6b5b2c2371eeb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290913
Reviewed-by: Travis Skare <skare@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#786994}