WebUI NTP: ntp-most-visited, ui fixes
The undo and restore buttons should only be shown in the toast when
an action is successful. This CL fixes a bug with data binding in the
HTML that was leading to the buttons always being shown.

When a user enters an invalid URL, this CL clears the error message
after the user changes the URL input.

Bug: 1069056, 1103803
Change-Id: I19096d83e9f7310dc389bf75b5cc0603a0efaedb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290544
Commit-Queue: Esmael Elmoslimany <aee@chromium.org>
Auto-Submit: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Tibor Goldschwendt <tiborg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787013}