media/gpu/vaapi: remove 'Task' from VaapiVideoDecoder method names
Usually methods have a "Task" suffix to indicate that they are run
on a given TaskRunner. VaapiVideoDecoder runs in its entirety on
its |decoder_sequence_checker_|, so there's no point in having this
suffix in a few methods, and can be confusing. This CL removes it.

No new functionality intended.

Bug: 1040291
Change-Id: Ie369f525696b874f3d252293ece7f996c51e91c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2289923
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: David Staessens <dstaessens@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787049}