[WebLayer] Reuse //chrome's backing Java infra for translate infobar
This CL eliminates //weblayer's backing Java infra for its translate
infobar implementation in favor of reusing //chrome's now that the
latter has been componentized. There was no substantive forking in any
of the files being eliminated and no behavioral change. The only change
necessary in the supporting files is to accommodate the fact that
strict mode violation detection is turned on in the context of WebLayer.

Bug: 1101240
Change-Id: I369741aef608a9227d6af12c361b13e8faba26a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279782
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787128}