Annotate steps in isolated-code-cache/same-origin-test.js
It was difficult to understand which steps in the test correspond to
which parts in the expectation. One particularly non-obvious part is
that producing the cache hits two trace events along the path, one of
which does not say it is responsible for producing the cache.

Proposing to output a comment message into the expectation before each
significant step performed by the test. Unfortunately, it leads to
flushing the timeline each time.

Bug: 1045052
Change-Id: I293b299fd1cd13d2ae938b43f7579730fc8f74c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290112
Reviewed-by: Mythri Alle <mythria@chromium.org>
Commit-Queue: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787142}