[Paint Preview] Implement hybrid approach for startup experiment
This CL adds a callback to PlayerManager for tracking when user
interactions with the player happen.

If the user doesn't interact with the paint preview before the live page
is loaded and ready, the paint preview will be removed and the live page
is displayed.

If the user interacts with the paint preview, it won't be removed.

Bug: 1064011
Change-Id: I1d0a56f11a6ec9ce899d294f4ad4b4e3c8e64fbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2288194
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Mehran Mahmoudi <mahmoudi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787209}