Remove disable-win32-lockdown flag.
Flag is no longer required (e.g. no reports in crash/).

Lockdown is only enabled if supported so it is not necessary to check
for support before enabling lockdown. However, some parts of the code
behave differently on Windows 7 where lockdown is not supported so
checks are now made on the platform version.

bug: 1059451
Change-Id: Ib3cd6311bfd5374c130f17c84ce9bdfb82512614
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2268538
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787288}