[Android R] Fix crash on tab restore with corrupt TabState.
In the case where a corrupted tab state is created of a certain form
(i.e. we map a 0-length byte buffer of contents state but don't hit EOF),
getting the capacity in native of the empty ByteBuffer returns -1
causing us to construct an invalid Pickle and crash. Previously this
would return 0 and would result in a failed tab restore / empty tab.

https://android-review.googlesource.com/c/platform/art/+/1276315/
changed this in R. Update Chrome and a regression test.

BUG=1094239

Change-Id: Ia4f3ac39ab74f4a8817fe41647ab5e9919f84198
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290055
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787292}