Link Doctor Removal Part 4: CaptivePortalBrowserTests.
The CaptivePortalBrowserTest fixture enables Link Doctor to make sure
that it works in combination with that feature. Link Doctor is slated
for removal, and disabled on all channels, so make those tests no longer
enable Link Doctor.

This has almost no impact on the tests - some of them no longer have to
wait for the double-commit-on-error that Link Doctor does, but that's
it.

Bug: 1030281
Change-Id: Ic0503d2a6ad83187795ea8e62b73774cbeda49d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2284059
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787301}