[AW] Fix Variations.SeedFreshness metric
crrev.com/c/2247154 introduced a regression that caused
SeedStore::RecordLastFetchTime to be given a null time rather than the
seed creation time because it attempted to read from the seed after it
had been std::moved. Variations.SeedFreshness is based on the last fetch
time, which is why it's be steadily climbing since that change landed.

Bug: 1103003
Change-Id: Id8bef58cc755473eb99fe6dc0090b998c5566492
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2285505
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787387}