Lazily load dynamic navigations for loading=lazy iframes
This CL introduces LazyLoadFrameObserver::LoadType enum, and the
LazyLoadFrameObserver is constructed with a member of this enum. This
member directs the frame observer to load a once-deferred frame in one
of two ways depending on whether or not this is the first navigation for
the frame, or dynamic / subsequent navigations.

R=kinuko@chromium.org, sclittle@chromium.org

Bug: 993273
Change-Id: I75cda5fc38ebfa351b1ad30a9b9ba78351f64a27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2289436
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Scott Little <sclittle@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787396}