overview: Animate the clip header while the window animates.
This was previously animated after the main animation for performance
reasons. With the improvements to overview, we think we run this
animation at the same time as the window animation, and add it to the
exit animation as well.

Since we now need to animate removing clipping (clip rect = gfx::Rect()),
this CL adds an observer which first animates the clip rect to the size
of the layer, and the removes the clip. Otherwise the animation will
contract hide everything before reshowing on the last frame. To help
with this, windows with no clip in overview will have a clip the size
of the window (visually equal to no clipping). This will avoid us
changing to and from real no clip (clip = gfx::Rect()) which would
make the logic more confusing and we would need to create more self
deleting observers.

Also refactors some logic so that this clipping, and the aspect ratio
clipping (clips windows that will open in splitview so they have the
same ratio as a splitview window), so they share more code.

Test: manual
Change-Id: I900f9a1801f9dd2e9847101e4c52d54f0ba41292
Fixed: 952158, 1103621
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2277331
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787412}