Put unit tests' anchor view in Widget to allow bubble creation
FeaturePromoBubbleView requires a ThemeProvider in its parent
Widget. So, for TabGroupsIPHControllerTest to instantiate it the
anchor view must be inside a Widget with a ThemeProvider.

TestWidget used internally to ChromeViewsTestBase serves this
purpose. This CL extracts it as ChromeTestWidget to be used
elsewhere. TabGroupsIPHControllerTest uses this as the parent Widget.

Bug: 1102655
Change-Id: Ib0d8a290789c1af1c08adb9fb6afc816ee2eeb90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2289922
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787419}