Adding UKMs for measuring beforeunload performance (1/2).
Breaking up this cl[1] into two portions: before unload and unload
metrics.

Adding UKMs to help measure unload performance.
More details on why we are adding these metrics can be found on the UKM
Collection Review doc[2]. TL;DR is that Search is looking to see where
they could improve unload and on the web platform side we are looking
to run experiments to prioritize unload and these metrics will help
measure impact.

This CL adds the metrics we want to collect for before unload. For
browser-initiated navigations we look at the queuing duration
(renderer-initiated navigations run before unload handlers synchronously
so we don't record queuing time. However, if the page has OOPIFs whose
beforeunloads will be dispatched from the browser, then we will record
queueing time for those) and for all navigations we
look at the duration of the before unload handlers (if it is a
renderer-initiated navigation who dispatches beforeunloads to OOPIFs
from the browser process, the duration of those will be summed into the
duration of the local beforeunloads). These metrics are
recorded for both cross-process and same-process navigations upon a
navigation commit and use the frames origin prior to the navigation
committing not the destination (ie if we navigate from google.com to
cnn.com the unload times are recorded for google.com).

Metrics added in this CL are:
1. BeforeUnloadDuration
2. BeforeUnloadQueueingDuration (measuring the time from browser sending the before unload to the unload being run in the renderer).

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2191143
[2] https://docs.google.com/document/d/1DECc8eJQluwkV5k64TX7Krl4zAcP7-Qk7pSirmHuCJs/edit?usp=sharing

Change-Id: Id6711f65ce3026c298b1b25d1b8a320b70adfdbe
Bug: 1081617
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2234546
Commit-Queue: Katie Dillon <kdillon@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Bryan McQuade <bmcquade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787427}