Switch VM permission service to use UnguessableToken
Even though GenerateGUID() is supposed to return cryptographically
strong data, it is better to use a dedicated implementation for the
token that is not supposed to be guessed by a 3rd party.

Note that we still use strings to exchange tokens, as we are using DBus
and not mojo (which would ensure type safety and would not expose
implementation details such as the fact that token is 2 64-bit numbers).

Bug: 1071872
Change-Id: I3014864201ee0af5e1d402dcf7ecb6c2e360f513
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2289213
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Daniel Ng <danielng@google.com>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787450}