Remove TestRunner reliance on the |main_view_| during test start and end
During startup, we want to disable the V8 cache if we're loading a
devtools test. That can be done directly from BlinkTestRunner where
the main frame is being loaded.

During test completion, we want to check the main frame's mime type
for text/plain, in order to change our output to text without a pixel
dump. This can be done directly against the main frame, since
BlinkTestRunner::TestFinished() runs against that main test window's
main frame.

This removes 2 more places causing problematic tracking of the "main
view" in the renderer, when it may change over the course of a test
due to frame swaps on cross-site navigations, and in the future on
same-site navigations as well.

R=avi@chromium.org

Bug: 866140, 1069111
Change-Id: I197601029cd02c227a26c93380db23ae35cb33d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2293375
Commit-Queue: danakj <danakj@chromium.org>
Auto-Submit: danakj <danakj@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787483}