Remove the TestRunner::main_view_ and the TestInterfaces class.
The last use of main_view_ was to get the main frame's current URL, but
grabbing the current frame instead, where queueLoad() is called from,
also works to build the desired relative URL for loading.

At that point TestInterfaces is doing nothing at all anymore, and can
be removed. There now exist no process-wide pointers down to track the
"main" RenderView/WebView/MainFrame.

From here, it should be relatively straight-forward to collapse the
BlinkTestRunner into WebViewTestProxy and/or TestRunner, depending on
if the code needs to be per-RenderView (such as is_main_window()?), or
if the code is part of the test control flow (such as TestFinished().

Then, it should hopefully again be much more straight-forward to move
code that is now process-global up to (single) the WebTestControlHost
in the browser process.

R=avi@chromium.org

Bug: 866140, 1069111
Change-Id: Iae7684e73f87d2203dce902be7efbc61cd13a891
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2292711
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787489}