Extensions: Bump DocServer version.
One of the files in the GCS webstore bucket doesn't seem to be updating.
Prima facie, this seems to be an issue with how we cache entities.
Speculatively bump the server version, as a short term fix.

Skipping presubmits since they seem to be timing out.

BUG=None
TBR=rdevlin.cronin@chromium.org

No-Presubmit: true
Change-Id: I1cee5ebeb1ee1b24c19203c97f15d877a352af94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2294093
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787560}