Fix missing a11y name in Ash security token PIN
Add the missing accessible name for the PinRequestView's input field
(i.e., the Textfield owned by FlexCodeInput).

Also fix the focus behavior to be only allowed for the input field
itself, not for its parent (i.e., the Textfield instead of the
FlexCodeInput). Without this fix, the a11y DCHECK complained that
there's a focusable control without an accessible name set.

This was caught thanks to the DCHECKs in the AccessibilityChecker.

Bug: 1001288
Test: run the browser_test from CL:2289411 with dchecks enabled
Change-Id: I8fa46004b8835f6d1d0afb19e5977ae61fc58240
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290151
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Fabian Sommer <fabiansommer@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787690}