heap, wtf: Concurrently mark VectorBackedLinkedList<Member<T>>
This CL sets kCanTraceConcurrently for VectorBackedLinkedListNode
according to its ValueType.

The CL also resolve a data race caused by moving element in the
vector. Moving an element uses the copy constructor. The race is
caused by writing to the value stored in the element (of type
VectorBackedLinkedListNode), which uses the value's copy constructor.
If that type is a Member<T>, the copy constructor uses non-atomic
writes. The new construct traits override this behavior for garbage
collected types and replaces it an atomic memcpy.

This is needed for concurrently marking NewLinkedHashSet.

Bug: 986235
Change-Id: I3758bcd4fe1090b8ef452711ea052113a2324be8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2292215
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787706}