Allow initial state determination after power wash of consumer devices
This CL treats a value of 0 for the check_enrollment RW VPD key as a
signal to do an initial state determination (a.k.a. initial enrollment
in the code right now). If the value is 1, it still does an FRE check.

This change allows for multiple initial state determinations that can
bring up the proper OOBE for license packaging (LP) as well as for
zero-touch enrollment (ZTE).

Writing check_enrollment=0 to the RW VPD is effectively made a no-op
now. We do not clean all the places in the code that write this value
in this CL in order to make it smaller. This can be done easily in a
followup CL.

See go/cros-oobe-state-determinations for design.

    --gtest_filter="*Wizard*WithInitialEnrollmentTest*"

(cherry picked from commit a37c8f9186a5929cf3db907ffa339d61f2c7ee1b)

Bug: chromium:1103006
Test: xvfb-run -s "-screen 0 800x600x24" out/Default/browser_tests \
Change-Id: Ia3f60cee26dbe0fc8c81818c04c483e50ac7aaaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2285517
Commit-Queue: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Roman Aleksandrov <raleksandrov@google.com>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#786873}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2294197
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/branch-heads/4183@{#463}
Cr-Branched-From: 740e9e8a40505392ba5c8e022a8024b3d018ca65-refs/heads/master@{#782793}