[CBCM] Move ProfileReportGenerator to components
Moves the core ProfileReportGenerator class from chrome/browser to
components so it can be reused by iOS. The desktop implementation of the
delegate is staying in chrome/browser.

For easier reviewing, the core class still takes the concrete delegate as
constructor argument, but in the following CL, this will be changed to use
the reporting delegate factory.

Bug: 1091936
Change-Id: I8f8ece003f24ba7c660f597562ae4448e737cc0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2288812
Reviewed-by: Owen Min <zmin@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#787810}