[Mirroring Service] migrate to libcast ANSWER messaging
This patch updates the mirroring service to use openscreen's Answer
message definition for mirror setting constraints.
This necessitates a migration from base::Value to
Json::Value as well, so this includes a refactor of the ReceiverResponse
class as well.

Other notes:
1. Previously, the ReceiverResponse struct didn't offer any type safety,
which is worrisome since it is essentially a std::variant without any
checking. This patch includes updates to the ReceiverResponse to ensure
it is used correctly by callers, instead of relying on them to check
everything.

2. Unit tests using the ReceiverResponse have been refactored to support
the above changes, including the attempts at implementing some
const-correctness.

3. Sender side letterboxing is now toggled from the ANSWER message.

4. The mirror settings have been updated to use the constraints from the
ANSWER message, using libcast's new capture recommendations.

Bug: 1002603, b/158030831, b/158032164,
Change-Id: I8f633e9c41d70fdcfcce9f8be697acc1838cc8e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2245897
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Yuri Wiitala <miu@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787821}