Re-enable ash unit test on scrollable shelf's presentation time
Still not sure why the original test case failed. One possible reason
is that the scroll distance in the previous code, which is 50, is not
enough to scroll the shelf to the end. However, on my local device,
it always works with the same gn args.

This CL implements a function to get the maximum scroll distance of
scrollable shelf. It ensures that the scroll distance in the test is
large enough.

Bug: 1100254
Change-Id: I99911faa6a51ada0fc827f740b019b2d20e18b41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2293468
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787837}