ambient: Implement UX spec on OS Settings page.
This patch implements the topic source picker for Ambient mode.
Basic functions:
(1) Clicking anywhere on an unselected row selects that row, except if
the click is on the subpage arrow.
(2) Clicking the subpage arrow on an unselected row leads to the subpage
but does not select the item.
(3) Clicking anywhere on an already-selected item leads users to the
subpage.
(4) Up/down arrows are used to move between items.

ss: https://screenshot.googleplex.com/vBKWPkWNZ5t

Bug: b/159746638
Test: Manual
Change-Id: Idd8efe31f8aadb7a690ecb8401fe434acc304596
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2284642
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Commit-Queue: Tao Wu <wutao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787853}