Make CrxDownloader a ref-counted thread safe type.
This avoid retention issues and allows instances of CrxDownloader to
be shared between sequences when interfacing with RPC modules such as
BITS/DO for Windows.

There are some mechanical changes such as removing macros.h and
replacing thread checkers with sequence checkers (but not for the
BITS downloader, which has thread affinity due to COM RPC).

Change-Id: I62e159d4ae8999ec70da94247f91f9cc26a5d278
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2293461
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787857}