Added async reveal to the Tab Search bubble
This change updates TabSearchBubbleView such that it is revealed
asynchronously after its WebContents has finished loading.

This helps in eliminating stuttering that currently occurs due to the
bubble being dynamically resized as the WebContents is loaded in.

Removed call to WasHidden() which causes issues with WebContents
resizing and is not needed for the TabSearch usecase.

Bug: 1099917
Change-Id: I5b5999b00535647fbccddb2b8c88ee1070030347
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264614
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787869}