[Tab Groups Collapse] Notify screen reader when group state toggles.
On Windows with NVDA this will say the state ("expanded" or "collapsed")
when the group header is toggled with keyboard.

Mac is currently not reading the state from the accessibility node data
so will need a follow-up CL to add the collapsed state of the group as
part of its title for Mac (and potentially ChromeOS and Linux).

Bug: 1018230
Change-Id: I262ba26c50bd30753ae865e0e8ef2095e784e58f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2296286
Commit-Queue: Charlene Yan <cyan@chromium.org>
Reviewed-by: Connie Wan <connily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#787985}