Don't skip already loaded custom fonts in FontFallbackIterator
When FontFallbackIterator iterates through a custom font, it checks
RangeSetContributesForHint() and if false, skips the custom font.
While the check was to introduced to prevent unnecesary loading of
fonts that don't load in new characters, it incorrectly blocks already
loaded custom fonts which can be used right away.

Therefore, this patch revises RangeSetContributesForHint() so that we
skip a font only when it's also a pending custom font.

Bug: 1101483
Change-Id: I04ed909d154b137b2bf45e712da199e6a5ddce7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278648
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788038}