Change CrostiniManager::LaunchContainerApplication to use CrostiniSuccessCallback
Renamed CrostiniLaunchAppCallback to CrostiniSuccessCallback and use it
for LaunchContainerApplication to make this function consistent with all
other functions which are passing failure_message in addition to the
bool success.  This will allow failure_reason messages from the backend
to be propagated.

Bug: 1092657
Change-Id: Ied8d3d00f5a3950d1751d689dbdb26cb9fb1c044
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2294741
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Auto-Submit: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788083}