Fix ChallengeResponseAuthKeysLoader shutdown crash
Fix the crash that could occur in ChallengeResponseAuthKeysLoader
in case the key loading operation didn't complete before the
browser shutdown. The crash was caused by this code trying to
obtain the already-deleted CertificateProviderService for the
being-deleted profile.

The fix is to observe the destruction of the sign-in profile and
to cancel all subsequent jobs after this.

Fixed: 1100058
Change-Id: I64dcac5f86a497bdd9934a14e2b66846a18dbd50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270450
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Fabian Sommer <fabiansommer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788094}