Make AXActionHandler inherit of a new class AXActionHandlerBase
Since we are migrating AXTreeIDs to use embedding tokens as their
backing store for frames (see https://crrev.com/c/2270065), we can
no longer can assume that frames will have a valid AXTreeID since
the time they're created, only after the first time that they have
finished loading a document (and will change for new loads).

For this reason we need to make changes to avoid RenderFrameHostImpl
from trying to self-assign an AXTreeID automatically on creation, as
it was the case until now due to RFHI inheriting from AXActionHandler,
which would automatically retrieve a new AXTreeID from the registry
and assign it as part of AXActionHandler's constructor.

In order to help with this, let's introduce a new base class that
will contain most of AXActionHandler's logic but won't automatically
set any valid AXTreeID on creation, and provide a SetAXTreeID() method
so that it can be set later on.

This new class (AXActionHandlerBase) will only need to be subclassed
directly by those classes that need finer control over the value of
the AXTreeID (e.g. RFHI after crrev.com/c/2270065 lands). If you don't
need that, you simply inherit from AXActionHandler (which is a subclass
of AXActionHandlerBase now) as before, which will automatically set
a valid AXTreeID on construction time as before.

Bug: 1094150
Change-Id: I75f126de571eaf472d54a2e05b11884c1f8af52a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2292520
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788115}